LayoutTests/fast/js/dfg-arguments-strict-mode-expected.txt

Tests what happens when you use arguments in strict mode and present the DFG with a tempting optimization opportunity.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f(1) is 1
PASS successfullyParsed is true

TEST COMPLETE
